Tracker Sensors Decision Making
Assuming the 4 line tracking sensors are in place, 2 in the middle and 1 on each side this algorithm can tell whether the vehicle is on the line, on a right / left turn, on a cross (right & left) or it is lost (all black). In the future we will use counters to 'memorize' the map so it does what we want.